+++ include/addressatcommand.h
@@ -26,10 +26,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SMARTREDIS_NONKEYEDCOMMAND_H
-#define SMARTREDIS_NONKEYEDCOMMAND_H
+#ifndef SMARTREDIS_ADDRESSATCOMMAND_H
+#define SMARTREDIS_ADDRESSATCOMMAND_H
 
-#include "command.h"
+#include "nonkeyedcommand.h"
 
 ///@file
 
@@ -38,54 +38,16 @@ namespace SmartRedis {
 class RedisServer;
 
 /*!
-*   \brief The NonKeyedCommand intermediary class constructs Client
-*          commands without keys. These commands use db node addresses.
-*   \details The KeyedCommand class has multiple methods for dealing
-*            with non-keyed commands.
+*   \brief The AddressAtCommand class constructs non-keyed Client
+*          commands that address the user given db node(s).
+*          This is a subclass of the NonKeyedCommand class.
 */
-class NonKeyedCommand : public Command
+class AddressAtCommand : public NonKeyedCommand
 {
     public:
-        /*!
-        *   \brief Set address and port for command
-        *          to be executed on
-        *   \param address Address of database
-        *   \param port Port of database
-        */
-        void set_exec_address_port(std::string address,
-                                   uint16_t port);
-
-        /*!
-        *   \brief Get address that command will be
-        *          to be executed on
-        *   \return std::string of address
-        *           if an address hasn't been set,
-        *                 returns an empty string
-        */
-        std::string get_address();
-
-        /*!
-        *   \brief Get port that command will be
-        *          to be executed on
-        *   \return uint16_t of port
-        *           if port hasn't been set, returns 0
-        */
-        uint16_t get_port();
-
-        virtual CommandReply runme(RedisServer * r) = 0;
-
-    private:
-        /*!
-        *   \brief Address of database node
-        */
-        std::string _address;
-
-        /*!
-        *   \brief Port of database node
-        */
-        uint64_t _port;
+        virtual CommandReply runme(RedisServer * r);
 };
 
 } //namespace SmartRedis
 
-#endif //NONKEYEDCOMMAND
+#endif //ADDRESSATCOMMAND
