+++ include/keyedcommand.h
@@ -58,78 +58,6 @@ class KeyedCommand : public Command
         virtual CommandReply runme(RedisServer * r);
 };
 
-/*!
-*   \brief The MultiKeyCommand class constructs Client
-*          commands with multiple keys. This is a subclass
-*          of the KeyedCommand class.
-*/
-class MultiKeyCommand : public KeyedCommand
-{
-    public:
-	    virtual CommandReply runme(RedisServer * r);
-};
-
-
-/*!
-*   \brief The SingleKeyCommand class constructs Client
-*          commands with only one key. This is a subclass
-*          of the KeyedCommand class.
-*/
-class SingleKeyCommand : public KeyedCommand
-{
-    public:
-	    virtual CommandReply runme(RedisServer * r);
-};
-
-/*!
-*   \brief The CompoundCommand class constructs Client
-*          commands with multiple commands. This is a subclass
-*          of the KeyedCommand class.
-*/
-class CompoundCommand : public KeyedCommand
-{
-    public:
-        virtual CommandReply runme(RedisServer * r);
-};
-
-
-class GetTensorCommand : public SingleKeyCommand
-{
-    public:
-        virtual CommandReply runme(RedisServer *r);
-
-        /*!
-        *   \brief This function will fill a vector with the dimensions of
-        *          the tensor.
-        *   \param reply CommandReply from running "AI.TENSORGET"
-        *   \return std::vecotr<size_t> with dimensions of the tensor
-        *   We assume right now that the META reply is always
-        *   in the same order and we can index base reply elements array;
-        */
-        std::vector<size_t> get_tensor_dims(CommandReply& reply);
-
-        /*!
-        *   \brief This function returns a string view of the data
-        *          tensor blob value
-        *   \param reply CommandReply from running "AI.TENSORGET"
-        *   \return string view of the data tensor blob value
-        *   We are going to assume right now that the meta data reply
-        *   is always in the same order and we are going to just
-        *   index into the base reply.
-        */
-        std::string_view get_tensor_data_blob(CommandReply& reply);
-
-        /*!
-        *   \brief This function returns a string of the tensor type
-        *   \param reply CommandReply from running "AI.TENSORGET"
-        *   \return string of the tensor type
-        *   We are going to assume right now that the meta data reply
-        *   is always in the same order and we are going to just
-        *   index into the base reply.
-        */
-       TensorType get_tensor_data_type(CommandReply& reply);
-};
-
 } //namespace SmartRedis
 
 #endif //KEYEDCOMMAND
