+++ src/cpp/nonkeyedcommand.cpp
@@ -48,95 +48,3 @@ uint16_t NonKeyedCommand::get_port()
 {
     return this -> _port;
 }
-
-CommandReply AddressAtCommand::runme(RedisServer * r)
-{
-    return r->run(*this);
-}
-
-CommandReply AddressAnyCommand::runme(RedisServer * r)
-{
-    return r->run(*this);
-}
-
-CommandReply DBInfoCommand::runme(RedisServer *r)
-{
-    return r->run(*this);
-}
-
-parsed_reply_nested_map DBInfoCommand::parse_db_node_info(std::string info)
-{
-    parsed_reply_nested_map info_map;
-
-    std::string delim = "\r\n";
-    std::string currKey = "";
-    size_t start = 0U;
-    size_t end = info.find(delim);
-
-    while (end != std::string::npos)
-    {
-        std::string line = info.substr(start, end-start);
-        start = end + delim.length();
-        end = info.find(delim, start);
-        if (line.length() == 0)
-            continue;
-        if (line[0] == '#')
-        {
-            currKey = line.substr(2);
-            if (info_map.find(currKey)==info_map.end())
-                info_map[currKey] = {};
-        }
-        else
-        {
-            std::size_t separatorIdx = line.find(':');
-            info_map[currKey][line.substr(0, separatorIdx)] =
-                                line.substr(separatorIdx+1);
-        }
-    }
-    std::string line = info.substr(start);
-    if (line.length() > 0)
-    {
-        std::size_t separatorIdx = line.find(':');
-        if (info_map.find(currKey)==info_map.end())
-                    info_map[currKey] = {};
-        info_map[currKey][line.substr(0, separatorIdx)] =
-                            line.substr(separatorIdx+1);
-    }
-    return info_map;
-}
-
-CommandReply ClusterInfoCommand::runme(RedisServer *r)
-{
-    return r->run(*this);
-}
-
-parsed_reply_map ClusterInfoCommand::parse_db_cluster_info(std::string info)
-{
-    parsed_reply_map info_map;
-
-    std::string delim = "\r\n";
-    std::string currKey = "";
-    size_t start = 0U;
-    size_t end = info.find(delim);
-
-    while (end != std::string::npos)
-    {
-        std::string line = info.substr(start, end-start);
-        start = end + delim.length();
-        end = info.find(delim, start);
-        if (line.length() == 0)
-            continue;
-
-        std::size_t separatorIdx = line.find(':');
-        info_map[line.substr(0, separatorIdx)] =
-                 line.substr(separatorIdx+1);
-    }
-    std::string line = info.substr(start);
-    if (line.length() > 0)
-    {
-        std::size_t separatorIdx = line.find(':');
-        info_map[line.substr(0, separatorIdx)] =
-                 line.substr(separatorIdx+1);
-    }
-    return info_map;
-}
