+++ src/cpp/singlekeycommand.cpp
@@ -26,13 +26,52 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "keyedcommand.h"
+#include "singlekeycommand.h"
 #include "redisserver.h"
 
 using namespace SmartRedis;
 
-CommandReply KeyedCommand::runme(RedisServer * r)
+CommandReply SingleKeyCommand::runme(RedisServer * r)
 {
-    CommandReply reply;
-    return reply;
+    return r->run(*this);
+}
+
+CommandReply GetTensorCommand::runme(RedisServer * r)
+{
+    return r->run(*this);
+}
+
+std::vector<size_t> GetTensorCommand::get_tensor_dims(CommandReply& reply)
+{
+    if(reply.n_elements() < 6)
+        throw std::runtime_error("The message does not have the "\
+                                "correct number of fields");
+
+    size_t n_dims = reply[3].n_elements();
+    std::vector<size_t> dims(n_dims);
+
+    for(size_t i=0; i<n_dims; i++) {
+        dims[i] = reply[3][i].integer();
+    }
+
+    return dims;
+}
+
+TensorType GetTensorCommand::get_tensor_data_type(CommandReply& reply)
+{
+    if(reply.n_elements() < 2)
+        throw std::runtime_error("The message does not have the correct "\
+                                "number of fields");
+
+    return TENSOR_TYPE_MAP.at(std::string(reply[1].str(),
+                                            reply[1].str_len()));
+}
+
+std::string_view GetTensorCommand::get_tensor_data_blob(CommandReply& reply)
+{
+    if(reply.n_elements() < 6)
+        throw std::runtime_error("The message does not have the "\
+                                "correct number of fields");
+
+    return std::string_view(reply[5].str(), reply[5].str_len());
 }
